refactor error downloading

+++ code/src/Core/Locations/RemoteTemplatesLocation.cs
@@ -16,139 +16,111 @@ namespace Microsoft.Templates.Core.Locations
         private readonly string CdnUrl = Configuration.Current.CdnUrl;
         private const string CdnTemplatesFileName = "Templates.mstx";
 
-        public override (LocationCopyStatus Status, string Message) Copy(string workingFolder)
+        public override void Copy(string workingFolder)
         {
-            return Download(workingFolder, CdnTemplatesFileName);
+            Download(workingFolder, CdnTemplatesFileName);
         }
 
         public override string GetVersion(string workingFolder)
         {
-            var version = "1.0.0";
+            var sourceUrl = $"{CdnUrl}/{VersionFileName}";
+            var destinationFile = Path.Combine(workingFolder, VersionFileName);
+
             try
             {
-                var sourceUrl = $"{CdnUrl}/{VersionFileName}";
-                var destinationFile = Path.Combine(workingFolder, VersionFileName);
-
                 var wc = new WebClient();
                 wc.DownloadFile(sourceUrl, destinationFile);
 
-                version = File.ReadAllText(destinationFile);
-
-                File.Delete(destinationFile);
+                var version = File.ReadAllText(destinationFile);
+                return version;
             }
             catch (Exception ex)
             {
                 AppHealth.Current.Error.TrackAsync("Error downloading the version file.",ex).FireAndForget();
+                throw;
             }
-
-            return version;
+            finally
+            {
+                if (File.Exists(destinationFile))
+                {
+                    File.Delete(destinationFile);
+                }
+            }
         }
 
-        private (LocationCopyStatus Status, string Message) Download(string workingFolder, string fileName)
+        private void Download(string workingFolder, string fileName)
         {
-            var status = LocationCopyStatus.Started;
-            var message = "Content adquisition started.";
-
             EnsureWorkingFolder(workingFolder);
 
             var sourceUrl = $"{CdnUrl}/{fileName}";
-            var tempFoder = Path.Combine(workingFolder, "temp");
+            var tempFolder = Path.Combine(workingFolder, "temp");
             var file = Path.Combine(workingFolder, fileName);
 
-           
-            if (DownloadContent(sourceUrl, file, out message))
-            {
-                status = LocationCopyStatus.SourceAdquired;
-                AppHealth.Current.Verbose.TrackAsync($"LocationCopyStatus: {status.ToString()}. {message}").FireAndForget();
+            DownloadContent(sourceUrl, file);
 
-                if (ExtractContent(file, tempFoder, out message))
-                {
-                    AppHealth.Current.Verbose.TrackAsync($"LocationCopyStatus: {status.ToString()}. {message}").FireAndForget();
-                    if (UpdateTemplatesFolder(tempFoder, workingFolder, out message))
-                    {
-                        status = LocationCopyStatus.TargetUpdated;
-                        AppHealth.Current.Verbose.TrackAsync($"LocationCopyStatus: {status.ToString()}. {message}").FireAndForget();
+            ExtractContent(file, tempFolder);
 
-                        if (CleanUpTempFiles(tempFoder, file, out message))
-                        {
-                            status = LocationCopyStatus.Finished;
-                            AppHealth.Current.Verbose.TrackAsync($"LocationCopyStatus: {status.ToString()}. {message}").FireAndForget();
-                        }
-                    }
-                }
-            }
-            return (status, message);
+            UpdateTemplates(tempFolder, workingFolder);
+
+            CleanUpTempFiles(tempFolder, file);
         }
 
-        private static bool ExtractContent(string file, string workingFolder, out string message)
+        private static void ExtractContent(string file, string workingFolder)
         {
             try
             {
                 Templatex.Extract(file, workingFolder);
-                message = "Content extracted successfully.";
-                return true;
             }
             catch(Exception ex)
             {
                 var msg = "The templates content can't be extracted.";
                 AppHealth.Current.Exception.TrackAsync(ex, msg).FireAndForget();
-                message = $"{msg}. Error message: {ex.Message}";
-                return false;
+                throw;
             }
         }
 
-        private static bool CleanUpTempFiles(string usedTempFolder, string mstxFile, out string message)
+        private static void CleanUpTempFiles(string usedTempFolder, string mstxFile)
         {
             try
             {
                 File.Delete(mstxFile);
                 Directory.Delete(usedTempFolder, true);
-                message = "Templates content updated successfully.";
-                return true;
             }
             catch(Exception ex)
             {
                 var msg = "The cleanup task can't be executed.";
                 AppHealth.Current.Exception.TrackAsync(ex, msg).FireAndForget();
-                message = $"{msg}. Error message: {ex.Message}";
-                return false;
+                throw;
             }
         }
 
-        private static bool UpdateTemplatesFolder(string tempFoder, string workingFolder, out string message)
+        private static void UpdateTemplates(string tempFoder, string workingFolder)
         {
             try
             {
                 SafeDelete(Path.Combine(workingFolder, TemplatesName));
                 CopyRecursive(Path.Combine(tempFoder, TemplatesName), Path.Combine(workingFolder, TemplatesName));
-                message = "Templates content updated successfully.";
-                return true;
             }
             catch(Exception ex)
             {
                 var msg = "The templates content can't be updated.";
                 AppHealth.Current.Exception.TrackAsync(ex, msg).FireAndForget();
-                message = $"{msg}. Error message: {ex.Message}";
-                return false;
+                throw;
             }
         }
 
-        private bool DownloadContent(string sourceUrl, string file, out string message)
+        private static void DownloadContent(string sourceUrl, string file)
         {
             try
             {
                 var wc = new WebClient();
                 wc.DownloadFile(sourceUrl, file);
-                message = "Templates content downloaded successfully.";
-                return true;
-
             }
             catch (Exception ex)
             {
                 string msg = "The templates content can't be downloaded.";
                 AppHealth.Current.Exception.TrackAsync(ex, msg).FireAndForget();
-                message = $"{msg}. Error message: {ex.Message}";
-                return false;
+                throw;
             }
         }
